Refactor email sending into shared utility
